wu-move-from-ipc-move: Move update_weights_from_ipc onto WeightUpdater (cut+paste)
Refactor chain ID: wu-move-from-ipc-move
